Clusterpath algorithm for Hüsler-Reiss models

Gradient descent based on the Clusterpath algorithm adapted to the likelihood of a graphical Hüsler-Reiss model. Usefull when the precision matrix \(\Theta\) (or the variogram \(\Gamma\)) has a block matrix structure.

Usage

HR_Clusterpath(
  data,
  lambda,
  zeta,
  W = NULL,
  p = NULL,
  eps_f = NULL,
  kappa = 0.01,
  eps_conv = 1e-07,
  tol_opt = 0.001,
  iter_max = 1000
)

Arguments

data

A \(n \times d\) matrix : the matrix of data.

lambda

If a positive number, the regularized parameter for the penalty. If a numerix vector of positive number : the grid line for \(\lambda\).

zeta

A positive number : tuned parameter for the exponential weights. If W is not NULL, this parameter is not used.

W

A \(d \times d\) matrix of positive number : the weights for the penalty. If NULL (default), it is set as the exponential weights, which depends of only one tuning parameter \(\zeta\) and where we have some theoretical results.

p

A numeric between 0 and 1, or NULL. If NULL (default), it is assumed that the data are already on multivariate Pareto scale. Else, p is used as the probability in the function data2mpareto() to standardize the data (see graphicalExtremes documentation).

eps_f

A positive number : tolerance threshold for merging clusters. If NULL (default), it is set as \(\kappa\) times the median of the non-diagonal elements of the distance matrix between the column of the initial \(\Theta\).

kappa

A positive number : tuned parameter for the fusion threshold. If eps_f is not NULL, this parameter is not used.

eps_conv

A positive number : tolerance threshold for the convergence of the algorithm.

iter_max

An integer : maximal number of iteration of the procedure.

Value

HR_Clusterpath() is a block gradient descent from the data with default values and method for the optimization : the variogram matrix \(\Gamma\) is the empirical variogram and the weights are set as the exponential weights, which depends of only one tuning parameter \(\zeta\) and where we have some theoretical results.

The produce a list of results :

$R : the \(R\) matrix of the clusters.

$clusters : a list of the variable indices, clustered.

$likelihood : the value of the negative penalised negative loglikelihood.

$lambda : the value of lambda.

$inputs : a list of the input parameters used in the procedure.

Details

The block matrix models is defined from two elements : the cluster's partition \(\{C_1, \dots, C_K\}\), included in \(V\) and the \(R\) matrix which belongs to \(\mathcal S_K(\mathbb R)\), the set of symmetric \(K \times K\) matrix (See also build_theta() and extract_R_matrix()).

The Clusterpath aims to find optimum of some penalised negative loglikelihood defined in neg_likelihood_pen().

Some results for HR_Clusterpath()

When we get replications \(X_1, \dots, X_n\) of a random vector \(X\) which belongs to the attraction domain of a \(K\)-block Hüsler-Reiss graphical model, the minimum of the penalised negative loglikelihood conveges almost surely to the true precision matrix of the model \(\Theta^*\), for all \(\lambda\), provided that we choose a sequence \((\zeta_n)_{n\in \mathbb N^*}\) which grows slower than the \(log(n)\) sequence. In general, the \(\zeta \in [1,2]\) is a good choice.

Examples

############################################################################
#                            With get_clusters
############################################################################
# Customizable weights
W <- matrix(c(0, 1, 1, 1,
              1, 0, 1, 1,
              1, 1, 0, 1,
              1, 1, 1, 0), nc = 4)

# Free choice of variogram
gamma <- graphicalExtremes::generate_random_Gamma(d = 4)

# Choice of initial condition for the optimization
R <- matrix(c(1, 0, 0, -1,
              0, 1, 1, -2,
              0, 1, 1, -1,
              -1, -2, -1, 1), nc = 4)
lambda <- 2.2

############################################################################
#                            With HR_Clusterpath
############################################################################
# Construction of clusters and R matrix for simulation
R <- matrix(c(1, -3, 0,
              -3, 2, -2,
              0, -2, 1), nc = 3)
clusters <- list(1:5, 6:10, 11:15)

# Construction of induced theta and corresponding variogram gamma
Theta <- build_Theta(R, clusters)
Gamma <- graphicalExtremes::Theta2Gamma(Theta)

gr3_bal_sim_param_cluster <-
  list(
    R = R,
    clusters = clusters,
    Theta = Theta,
    Gamma = Gamma,
    chi = 1,
    n = 1e3,
    d = 15
  )