Hüsler-Reiss Clusterpath refitted

The refit procedure is based on a convex optimization problem which is solved with the CVXR package. Based on the results produced by HR_Clusterpath, a new problem is defined as the minimization of the negative loglikelihood of a Hüsler-Reiss model with precision matrix \(\Theta\) under the constraints of block matrix structure induced by the estimated clusters. These solutions are computed for each value of \(\lambda\) in the output of HR_Clusterpath.

Usage

HR_Clusterpath_refit(hrc_output)

Arguments

hrc_output

An object of the HR_Clusterpath function.

Value

An object of class HR_Clusterpath_refit with the same structure as the HR_Clusterpath output but with refitted R matrices for each result.

Motivation

The problem solution provided by HR_clusterpath may be biased due to the penalty in the negative loglikelihood. To avoid this potential issue, the matrix \(R\) is refitted for each solution over \(\lambda\) with the non-penalized negative loglikelihood constraining by the estimated clusters.

Examples

# Initialization of the parameters for the simulation
R <- matrix(c(1, -3, 0,
              -3, 2, -2,
              0, -2, 1), nc = 3)
clusters <- list(1:5, 6:10, 11:15)

# Building of the variogram matrix for the simulation
Theta <- build_Theta(R, clusters)
Gamma <- graphicalExtremes::Theta2Gamma(Theta)

# Simulation of data
set.seed(123)
n <- 1e4
d <- ncol(Gamma)
data <- graphicalExtremes::rmstable(
           n = n,
           model = "HR",
           d = d,
           par = Gamma)

# Computation of the multivariate Pareto data
norm_inf <- apply(abs(data), 1, max)
quantile <- sort(norm_inf, decreasing = TRUE) |> as.numeric()
k <- floor(0.1 * nrow(data))
u <- quantile[k]

data_par <- data[norm_inf > u, ] / u         # Multivariate Pareto data

# Computation of the Hüsler-Reiss Clusterpath
zeta <- log(n) ** 2              # Zeta parameter for the exponential weights
HRC <- HR_Clusterpath(data_par, lambda = 10, zeta = zeta)

# First estimation of R
HRC$results$l1$R
#>             [,1]      [,2]        [,3]
#> [1,]  1.18671702 -3.153842 -0.09860223
#> [2,] -3.15384155  1.824990 -1.85374153
#> [3,] -0.09860223 -1.853742  0.90719632

# Refit of the Hüsler-Reiss Clusterpath
HRC_refit <- HR_Clusterpath_refit(HRC)

# New estimation of R
HRC_refit$results$l1$R
#>            [,1]      [,2]       [,3]
#> [1,]  1.0695803 -3.002448 -0.1067049
#> [2,] -3.0024481  1.562946 -1.8063244
#> [3,] -0.1067049 -1.806324  0.8696769