Clusterpath algorithm for Hüsler-Reiss models hierarchical version

Block gradient descent based on the Clusterpath algorithm for Hüsler-Reiss graphical model. Usefull when the precision matrix \(\Theta\) (or the variogram \(\Gamma\)) has a block matrix structure.

Usage

HR_Clusterpath_hierarchy(
  Gamma,
  lambda,
  W = NULL,
  zeta,
  eps_f = NULL,
  kappa = 0.01,
  eps_conv = 1e-07,
  tol_opt = 0.001,
  iter_max = 1000
)

Arguments

Gamma

A \(d \times d\) matrix : a variogram matrix.

lambda

A numeric vector of positive number : the grid line for \(\lambda\).

W

A \(d \times d\) matrix of positive number : the weights for the penalty. If NULL (default), it is set as the exponential weights, which depends of only one tuning parameter \(\zeta\) and where we have some theoretical results.

zeta

A positive number : tuned parameter for the exponential weights. If W is not NULL, this parameter is not used.

eps_f

A positive number : tolerance threshold for merging clusters. If NULL (default), it is set as \(\kappa\) times the median of the non-diagonal elements of the distance matrix between the column of the initial \(\Theta\).

kappa

A positive number : tuned parameter for the fusion threshold. If eps_f is not NULL, this parameter is not used.

eps_conv

A positive number : tolerance threshold for the convergence of the algorithm.

tol_opt

A positive number : tolerance for the optimal step in the gradient descent.

iter_max

An integer : maximal number of iterations of the procedure.

Value

HR_Clusterpath() is a block gradient descent from the data with default values and method for the optimization : the variogram matrix \(\Gamma\) is the empirical variogram and the weights are set as the exponential weights, which depends of only one tuning parameter \(\zeta\) and where some theoretical guarantees are provided.

The output is an object of class HRC_hierarchy which contains a list of several datas :

$results : a list of results for each value of lambda, where each result is a list with the following elements : $R the \(R\) matrix of the clusters, $clusters a list of the variable indices separated per cluster, $likelihood the value of the penalised negative loglikelihood and $lambda the corresponding regularized parameter.

$Gamma : the variogram matrix used for the procedure.

$inputs : a list of the input parameters used in the procedure.

Details

The block matrix models are characterized by two objects : a clustering partition \(\{C_1, \dots, C_K\}\) of \(V = \{1, \dots, d\}\) and a matrix \(R\) which belongs to \(\mathcal S_K(\mathbb R)\), the set of symmetric \(K \times K\) matrix (See also build_theta() and extract_R_matrix()).

The Clusterpath aims to find the optimum of some penalised negative loglikelihood defined in neg_likelihood_pen(). The path of solution is computed using the Clusterpath procedure on a mesh for \(\lambda\) where each initialization is provided using the results of the previous \(\lambda\) within the grid.

Examples

# Initialization of the parameters for the simulation
R <- matrix(c(1, -3, 0,
              -3, 2, -2,
              0, -2, 1), nc = 3)
clusters <- list(1:5, 6:10, 11:15)

# Building of the variogram matrix for the simulation
Theta <- build_Theta(R, clusters)
Gamma <- graphicalExtremes::Theta2Gamma(Theta)

# Simulation of data
set.seed(123)
n <- 1e4                               # Sample size
d <- ncol(Gamma)                       # Number of variables
data <- graphicalExtremes::rmstable(
           n = n,
           model = "HR",
           d = d,
           par = Gamma)

# Computation of the multivariate Pareto data
norm_inf <- apply(abs(data), 1, max)
quantile <- sort(norm_inf, decreasing = TRUE) |> as.numeric()
k <- floor(0.1 * nrow(data))
u <- quantile[k]

data_par <- data[norm_inf > u, ] / u        # Multivariate Pareto data

# Computation of the Hüsler-Reiss Clusterpath
zeta <- log(n) ** 2              # Zeta parameter for the exponential weights
Gamma_est <- graphicalExtremes::emp_vario(data_par) # Estimated variogram
lambda <- seq(0, 1, 0.1)                            # Meshgrid for lambda

HRC <- HR_Clusterpath_hierarchy(Gamma_est, lambda = lambda, zeta = zeta)
HRC$results$l10
#> $R
#>             [,1]      [,2]       [,3]        [,4]       [,5]
#> [1,]  1.17748338 -3.157533 -0.0463839 -0.07416062 -0.1288338
#> [2,] -3.15753309  1.678379 -1.8112093 -1.90786531 -1.9802285
#> [3,] -0.04638390 -1.811209  0.6917326  0.90379592  1.0108776
#> [4,] -0.07416062 -1.907865  0.9037959  0.94368511  1.0417123
#> [5,] -0.12883384 -1.980229  1.0108776  1.04171228  6.7747802
#> 
#> $clusters
#> $clusters$C1
#> [1] 1 3 5 2 4
#> 
#> $clusters$C2
#> [1]  6  8  9 10  7
#> 
#> $clusters$C3
#> [1] 11 12
#> 
#> $clusters$C4
#> [1] 13 15
#> 
#> $clusters$C5
#> [1] 14
#> 
#> 
#> $likelihood
#> [1] -19.13905
#> 
#> $lambda
#> [1] 0.9
#>