Gradient of the negative log likelihood without penalty

Gradient of the negative log likelihood without penalty

Usage

nloglike_grad_np(gamma)

Arguments

gamma

a d x d variogram matrix.

Value

A function of the R matrix and clusters and compute the gradient matrix of the negative log likelihood for a fixed variogram gamma. The gradient matrix can be computed by :

                   dnllh = dlog + dtrace

where : - dlog = t(U) g(Theta_p) U - 0.5 diag(t(U) g(Theta_p) U) - dtrace = - t(U) gamma U + 0.5 diag(t(U) gamma U)

Examples

R <- matrix(c(0.5, -1,
              -1, -1), nr = 2)
clusters <- list(c(1,3), c(2,4))
gamma <- matrix(c(0,2,1,0,
                  2,0,4,1,
                  1,4,0,7,
                  0,1,7,0), nc = 4)
gradient <- nloglike_grad_np(gamma)
#> Error in nloglike_grad_np(gamma): could not find function "nloglike_grad_np"
gradient(R, clusters)
#> Error in gradient(R, clusters): could not find function "gradient"