Clusterpath algorithm for Hüsler-Reiss models

Gradient descent based on the Clusterpath algorithm adapted to the likelihood of a graphical Hüsler-Reiss model. Usefull when the precision matrix \(\Theta\) (or the variogram \(\Gamma\)) has a block matrix structure.

Usage

get_cluster(gamma, weights, eps_f, ...)

HR_Clusterpath(data, zeta, lambda, eps_g = 0.001, eps_f = 0.01, it_max = 1000)

Arguments

gamma

For get_clusters(), a \(d \times d\) matrix : the variogram matrix \(\Gamma\).

weights

For get_clusters(), the \(d \times d\) symmetric weightsmatrix with a zero diagonal.

eps_f

A positive number : tolerance threshold for merging clusters.

data

For HR_Clusterapth(), a \(n \times d\) matrix : the matrix of data.

zeta

For HR_Clusterapth(), a positive number : tuned parameter for the exponential weights.

lambda

For get_clusters(), a positive number : the weight of the penalty.

For HR_Clusterpath(), a numerix vector of positive number : the grid line for \(\lambda\).

eps_g

A positive number : tolerance threshold for the convergence of the gradient descent.

it_max

An integer : maximal number of iteration of the gradient descent algorithm.

Value

HR_Clusterpath() is a gradient descent from the data with default values and method for the optimization : the variogram matrix \(\Gamma\) is the empirical variogram and the weights are set as the exponential weights, which depends of only one tuning parameter \(\zeta\) and where we have some theoretical results.

get_clusters() is a freer version of the previous function. You can use your own estimation for the variogram \(\Gamma\) and customizable weights with the matrix \(W\).

Both of them produce a list of results :

$R : the \(R\) matrix of the clusters.

$clusters : a list of the variable indices, clustered.

$nllh : the value of the negative penalised negative loglikelihood.

$lambda : the value of lambda.

In the case of HR_Clusterpath() it is a list of the previous list.

Details

The block matrix models is defined from two elements : the cluster's partition \(\{C_1, \dots, C_K\}\), included in \(V\) and the \(R\) matrix which belongs to \(\mathcal S_K(\mathbb R)\), the set of symmetric \(K \times K\) matrix (See also build_theta() and extract_R_matrix()).

The Clusterpath aims to find optimum of some penalised negative loglikelihood defined in neg_likelihood_pen().

Some results for HR_Clusterpath()

When we get replications \(X_1, \dots, X_n\) of a random vector \(X\) which belongs to the attraction domain of a \(K\)-block Hüsler-Reiss graphical model, the minimum of the penalised negative loglikelihood conveges almost surely to the true precision matrix of the model \(\Theta^*\), for all \(\lambda\), provided that we choose a sequence \((\zeta_n)_{n\in \mathbb N^*}\) which grows slower than the \(log(n)\) sequence. In general, the \(\zeta \in [1,2]\) is a good choice.

Examples

W <- matrix(c(0, 1, 1, 1,
              1, 0, 1, 1,
              1, 1, 0, 1,
              1, 1, 1, 0), nc = 4)
gamma <- generate_random_Gamma(d = 4)
#> Error in generate_random_Gamma(d = 4): could not find function "generate_random_Gamma"
R <- matrix(c(1,0,0,-1,
              0,1,1,-2,
              0,1,1,-1,
              -1,-2,-1,1), nc = 4)
Cluster_HR <- get_cluster(gamma, W, 100)
Cluster_HR(R)
#> Error in Cluster_HR(R): argument "lambda" is missing, with no default