Clusterpath algorithm for Hüsler-Reiss models

Gradient descent based on the Clusterpath algorithm adapted to the likelihood of a graphical Hüsler-Reiss model. Usefull when the precision matrix \(\Theta\) (or the variogram \(\Gamma\)) has a block matrix structure.

Usage

HR_Clusterpath(
  data,
  lambda,
  zeta,
  W = NULL,
  p = NULL,
  eps_f = NULL,
  kappa = 0.01,
  eps_conv = 1e-07,
  tol_opt = 0.001,
  iter_max = 1000
)

Arguments

data

A \(n \times d\) matrix : the matrix of data.

lambda

If a positive number, the regularized parameter for the penalty. If a numerix vector of positive number : the grid line for \(\lambda\).

zeta

A positive number : tuned parameter for the exponential weights. If W is not NULL, this parameter is not used.

W

A \(d \times d\) matrix of positive number : the weights for the penalty. If NULL (default), it is set as the exponential weights, which depends of only one tuning parameter \(\zeta\) and where we have some theoretical results.

p

A numeric between 0 and 1, or NULL. If NULL (default), it is assumed that the data are already on multivariate Pareto scale. Else, p is used as the probability in the function data2mpareto() to standardize the data (see graphicalExtremes documentation).

eps_f

A positive number : tolerance threshold for merging clusters. If NULL (default), it is set as \(\kappa\) times the median of the non-diagonal elements of the distance matrix between the column of the initial \(\Theta\).

kappa

A positive number : tuned parameter for the fusion threshold. If eps_f is not NULL, this parameter is not used.

eps_conv

A positive number : tolerance threshold for the convergence of the algorithm.

iter_max

An integer : maximal number of iteration of the procedure.

Value

HR_Clusterpath() is a block gradient descent from the data with default values and method for the optimization : the variogram matrix \(\Gamma\) is the empirical variogram and the weights are set as the exponential weights, which depends of only one tuning parameter \(\zeta\) and where we have some theoretical results.

The produce a list of several datas :

$results : a list of results for each value of lambda, where each result is a list with the following elements : $R the \(R\) matrix of the clusters, $clusters a list of the variable indices separated per cluster, $likelihood the value of the penalised negative loglikelihood and $lambda the corresponding regularized parameter.

$Gamma : the variogram matrix used for the procedure.

$inputs : a list of the input parameters used in the procedure.

Details

The block matrix models is defined from two elements : the cluster's partition \(\{C_1, \dots, C_K\}\), included in \(V\) and the \(R\) matrix which belongs to \(\mathcal S_K(\mathbb R)\), the set of symmetric \(K \times K\) matrix (See also build_theta() and extract_R_matrix()).

The Clusterpath aims to find optimum of some penalised negative loglikelihood defined in neg_likelihood_pen().

Some results for HR_Clusterpath()

When we get replications \(X_1, \dots, X_n\) of a random vector \(X\) which belongs to the attraction domain of a \(K\)-block Hüsler-Reiss graphical model, the minimum of the penalised negative loglikelihood conveges almost surely to the true precision matrix of the model \(\Theta^*\), for all \(\lambda\), provided that we choose a sequence \((\zeta_n)_{n\in \mathbb N^*}\) which grows slower than the \(log(n)^s\) sequence with \(s < 8\).

Examples

############################################################################
#                 EXAMPLE OF USE OF THE HR_CLUSTERPATH FUNCTION
############################################################################
# Construction of clusters and R matrix for simulation
R <- matrix(c(1, -3, 0,
              -3, 2, -2,
              0, -2, 1), nc = 3)
clusters <- list(1:5, 6:10, 11:15)

# Construction of induced theta and corresponding variogram gamma
Theta <- build_Theta(R, clusters)
Gamma <- graphicalExtremes::Theta2Gamma(Theta)

gr3_bal_sim_param_cluster <-
  list(
    R = R,
    clusters = clusters,
    Theta = Theta,
    Gamma = Gamma,
    n = 1e4,
    d = 15
  )

# Simulation of data
set.seed(123)
data <- graphicalExtremes::rmstable(
           n = gr3_bal_sim_param_cluster$n,
           model = "HR",
           d = gr3_bal_sim_param_cluster$d,
           par = gr3_bal_sim_param_cluster$Gamma)
norm_inf <- apply(abs(data), 1, max)
quantile <- sort(norm_inf, decreasing = TRUE) |> as.numeric()

k <- floor(0.1 * nrow(data))
u <- quantile[k]

data_par <- data[norm_inf > u, ] / u # Transformation to multivariate Pareto scale

# Computation of the solution produced by the clusterpath procedure for one value of lambda
zeta <- log(gr3_bal_sim_param_cluster$n) ** 2              # Zeta parameter for the exponential weights
HR_Clusterpath(data_par, lambda = 10, zeta = zeta)
#> $results
#> $results$l1
#> $results$l1$R
#>             [,1]      [,2]        [,3]
#> [1,]  1.18671702 -3.153842 -0.09860223
#> [2,] -3.15384155  1.824990 -1.85374153
#> [3,] -0.09860223 -1.853742  0.90719632
#> 
#> $results$l1$clusters
#> $results$l1$clusters$C1
#> [1] 1 3 5 2 4
#> 
#> $results$l1$clusters$C2
#> [1]  6  7  8  9 10
#> 
#> $results$l1$clusters$C3
#> [1] 11 12 13 15 14
#> 
#> 
#> $results$l1$likelihood
#> [1] -18.99939
#> 
#> $results$l1$lambda
#> [1] 10
#> 
#> 
#> 
#> $Gamma
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.0000000 0.1935635 0.1976865 0.1925554 0.1942073 0.1305479 0.1423761
#>  [2,] 0.1935635 0.0000000 0.1923807 0.1935749 0.1953140 0.1285176 0.1388852
#>  [3,] 0.1976865 0.1923807 0.0000000 0.2028182 0.2021537 0.1386159 0.1434535
#>  [4,] 0.1925554 0.1935749 0.2028182 0.0000000 0.2006503 0.1523259 0.1378902
#>  [5,] 0.1942073 0.1953140 0.2021537 0.2006503 0.0000000 0.1386372 0.1371836
#>  [6,] 0.1305479 0.1285176 0.1386159 0.1523259 0.1386372 0.0000000 0.1165319
#>  [7,] 0.1423761 0.1388852 0.1434535 0.1378902 0.1371836 0.1165319 0.0000000
#>  [8,] 0.1431203 0.1371816 0.1388117 0.1422908 0.1471640 0.1233934 0.1169895
#>  [9,] 0.1478524 0.1428143 0.1439434 0.1402170 0.1489401 0.1274285 0.1254620
#> [10,] 0.1397738 0.1474117 0.1324602 0.1363570 0.1381069 0.1200365 0.1253890
#> [11,] 0.2694862 0.2796950 0.2741213 0.2681132 0.2788989 0.2256845 0.2449816
#> [12,] 0.2823563 0.2648515 0.2625774 0.2649961 0.2892359 0.2291209 0.2401148
#> [13,] 0.2626023 0.2744026 0.2491146 0.2698384 0.2468328 0.2159792 0.2228590
#> [14,] 0.2534999 0.2585254 0.2500331 0.2522496 0.2449262 0.2118019 0.2025298
#> [15,] 0.2545866 0.2675285 0.2615758 0.2728267 0.2582310 0.2242242 0.2160183
#>            [,8]      [,9]     [,10]     [,11]     [,12]     [,13]     [,14]
#>  [1,] 0.1431203 0.1478524 0.1397738 0.2694862 0.2823563 0.2626023 0.2534999
#>  [2,] 0.1371816 0.1428143 0.1474117 0.2796950 0.2648515 0.2744026 0.2585254
#>  [3,] 0.1388117 0.1439434 0.1324602 0.2741213 0.2625774 0.2491146 0.2500331
#>  [4,] 0.1422908 0.1402170 0.1363570 0.2681132 0.2649961 0.2698384 0.2522496
#>  [5,] 0.1471640 0.1489401 0.1381069 0.2788989 0.2892359 0.2468328 0.2449262
#>  [6,] 0.1233934 0.1274285 0.1200365 0.2256845 0.2291209 0.2159792 0.2118019
#>  [7,] 0.1169895 0.1254620 0.1253890 0.2449816 0.2401148 0.2228590 0.2025298
#>  [8,] 0.0000000 0.1315627 0.1230037 0.2197020 0.2255586 0.2325263 0.2161046
#>  [9,] 0.1315627 0.0000000 0.1280559 0.2294147 0.2336325 0.2169210 0.2159390
#> [10,] 0.1230037 0.1280559 0.0000000 0.2380463 0.2347555 0.2141257 0.2146581
#> [11,] 0.2197020 0.2294147 0.2380463 0.0000000 0.3867288 0.3740374 0.3869221
#> [12,] 0.2255586 0.2336325 0.2347555 0.3867288 0.0000000 0.3869388 0.3975203
#> [13,] 0.2325263 0.2169210 0.2141257 0.3740374 0.3869388 0.0000000 0.3793816
#> [14,] 0.2161046 0.2159390 0.2146581 0.3869221 0.3975203 0.3793816 0.0000000
#> [15,] 0.2180682 0.2227965 0.2268491 0.3933264 0.3918188 0.3754891 0.3635154
#>           [,15]
#>  [1,] 0.2545866
#>  [2,] 0.2675285
#>  [3,] 0.2615758
#>  [4,] 0.2728267
#>  [5,] 0.2582310
#>  [6,] 0.2242242
#>  [7,] 0.2160183
#>  [8,] 0.2180682
#>  [9,] 0.2227965
#> [10,] 0.2268491
#> [11,] 0.3933264
#> [12,] 0.3918188
#> [13,] 0.3754891
#> [14,] 0.3635154
#> [15,] 0.0000000
#> 
#> $inputs
#> $inputs$eps_conv
#> [1] 1e-07
#> 
#> $inputs$tol_opt
#> [1] 0.001
#> 
#> $inputs$iter_max
#> [1] 1000
#> 
#> 
#> attr(,"class")
#> [1] "HR_Clusterpath"